requirements.txt
celery
flask
flask-socketio
python-libnmap
redis
